pyproject.toml
[tool.poetry]
name = "python-bhtmeter"
version = "0.1.0"
description = "Baro-, Hygro- and Thermo-meter measurement hub"
authors = ["Tadashi G. Takaoka <tadashi.g.takaoka@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.2"
tomli = "^2.0.1"
importlib = "^1.0.4"
bluepy = "^1.3.0"
pimoroni-bme280 = "^0.1.1"
ambient = {git = "https://github.com/TakehikoShimojima/ambient-python-lib.git"}
smbus2 = "^0.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests",]
